+++ erdpy/utils.py
@@ -8,12 +8,15 @@ import stat
 import sys
 import tarfile
 import zipfile
+import toml
+import binascii
+
 from pathlib import Path
 from typing import Any, Dict, List, Union
-
-import toml
-
-from erdpy import errors
+from os import path
+from erdpy import errors, guards
+from erdpy.accounts import Address
+from erdpy.errors import CannotReadValidatorsData
 
 logger = logging.getLogger("utils")
 
@@ -192,3 +195,4 @@ def breakpoint():
     print("Waiting for debugger attach")
     debugpy.wait_for_client()
     debugpy.breakpoint()
+